+++ src/components/charger-overview-card.ts
@@ -11,6 +11,7 @@ import { ChargerStore } from "../state/store";
 import "./charger-setpoint-controls";
 import "./charger-power-control";
 import "./charger-status-preview";
+import "./cooling-status";
 import "./rectifier-details";
 
 export const CHARGER_OVERVIEW_CARD_TAG = "r4875g1-charger-overview-card";
@@ -92,7 +93,7 @@ export class ChargerOverviewCard extends LitElement {
   }
 
   getCardSize(): number {
-    return 7;
+    return 9;
   }
 
   connectedCallback(): void {
@@ -130,6 +131,9 @@ export class ChargerOverviewCard extends LitElement {
           .store=${this.chargerStore}
           .execute=${this.executeControl}
         ></r4875g1-rectifier-details>
+        <r4875g1-cooling-status
+          .store=${this.chargerStore}
+        ></r4875g1-cooling-status>
         <r4875g1-charger-setpoint-controls
           .store=${this.chargerStore}
           .execute=${this.executeControl}